+++ modules/Content/assets/js/content.js
@@ -2,5 +2,4 @@
 
 // Global Vue components
 VueView.component('field-contentItemtLink', 'content:assets/vue-components/field-content-item-link.js');
-VueView.component('fields-manager', 'system:assets/vue-components/fields-manager.js');
 VueView.component('options-linkModel', 'content:assets/vue-components/options-link-model.js');