+++ modules/Content/views/collection/items.php
@@ -10,7 +10,7 @@
 
             <div class="kiss-flex kiss-flex-middle kiss-margin-bottom">
                 <div class="kiss-margin-small-right">
-                    <kiss-svg class="kiss-margin-auto" src="<?= $this->base('content:assets/icons/collection.svg') ?>" width="25" height="25" style="color:<?= ($this->escape($model['color'] ?? 'inherit')) ?>"><canvas width="35" height="35"></canvas></kiss-svg>
+                    <kiss-svg class="kiss-margin-auto" src="<?= $this->base(isset($model['icon']) && $model['icon'] ? $model['icon'] : 'content:assets/icons/collection.svg') ?>" width="30" height="30" style="color:<?= ($this->escape($model['color'] ?? 'inherit')) ?>"><canvas width="35" height="35"></canvas></kiss-svg>
                 </div>
                 <div class="kiss-margin-small-right">
                     <div class="kiss-size-5 kiss-text-bold"><?= $this->escape($model['label'] ? $model['label'] : $model['name']) ?></div>